Moved out of the __init__.py the functions: get_pokemon_id_by_tier()
