+++ src/Ankimon/functions/pokemon_functions.py
@@ -8,7 +8,17 @@ from aqt.utils import showWarning
 
 from .pokedex_functions import search_pokeapi_db_by_id, search_pokedex, search_pokedex_by_id, get_all_pokemon_moves
 from .battle_functions import calculate_hp
-from ..resources import pokedex_path, next_lvl_file_path, mypokemon_path, learnset_path
+from ..resources import (
+    pokedex_path,
+    next_lvl_file_path,
+    mypokemon_path,
+    learnset_path,
+    pokemon_species_baby_path,
+    pokemon_species_legendary_path,
+    pokemon_species_mythical_path,
+    pokemon_species_normal_path,
+    pokemon_species_ultra_path,
+)
 
 def pick_random_gender(pokemon_name):
     """
@@ -305,3 +315,23 @@ def check_min_generate_level(name):
     else:
         min_level = 1
         return min_level
+    
+def get_pokemon_id_by_tier(tier):
+    id_species_path = None
+    if tier == "Normal":
+        id_species_path = pokemon_species_normal_path
+    elif tier == "Baby":
+        id_species_path = pokemon_species_baby_path
+    elif tier == "Ultra":
+        id_species_path = pokemon_species_ultra_path
+    elif tier == "Legendary":
+        id_species_path = pokemon_species_legendary_path
+    elif tier == "Mythical":
+        id_species_path = pokemon_species_mythical_path
+
+    with open(id_species_path, "r", encoding="utf-8") as file:
+        id_data = json.load(file)
+
+    # Select a random Pokemon ID from those in the tier
+    random_pokemon_id = random.choice(id_data)
+    return random_pokemon_id
